game-music-emu: Merge kode54's Game Music Emu fork
From https://github.com/wyatt8740/Game_Music_Emu

With fixes:
- retain VGZ (gzip magic header) detection in gme_open_data (gme.cpp)
- fix some new Clang warnings

+++ game-music-emu/gme/Data_Reader.h
@@ -134,7 +134,8 @@ private:
 // Invokes callback function to read data. Size of data must be specified in advance.
 class Callback_Reader : public Data_Reader {
 public:
-	typedef const char* (*callback_t)( void* data, void* out, int count );
+/*	typedef const char* (*callback_t)( void* data, void* out, int count );*/
+	typedef const char* (*callback_t)( void* data, void* out, long count );
 	Callback_Reader( callback_t, long size, void* data = 0 );
 public:
 	long read_avail( void*, long );
